+++ lib/BlockchainDB/Lmdb/db_lmdb.h
@@ -229,7 +229,7 @@ public:
 
   virtual output_data_t get_output_key(const uint64_t& amount, const uint32_t& index);
   virtual output_data_t get_output_key(const uint32_t& global_index) const;
-  virtual void get_output_key(const uint64_t &amount, const std::vector<uint64_t> &offsets, std::vector<output_data_t> &outputs, bool allow_partial = false);
+  virtual void get_output_key(const uint64_t &amount, const std::vector<uint32_t> &offsets, std::vector<output_data_t> &outputs, bool allow_partial = false);
 
   virtual tx_out_index get_output_tx_and_index_from_global(const uint64_t& index) const;
   virtual void get_output_tx_and_index_from_global(const std::vector<uint64_t> &global_indices,
